tests/requirements.txt
tox
coverage
nose
nose-exclude
django-nose